make LoginController starting point of app, update voting.fxml
