+++ src/controller/LoginController.java
@@ -1,5 +1,11 @@
 package controller;
 
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXTextField;
 import javafx.fxml.FXML;
@@ -7,7 +13,7 @@ import javafx.fxml.FXML;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class LoginController {
+public class LoginController extends Application{
 
     @FXML
     private ResourceBundle resources;
@@ -24,5 +30,21 @@ public class LoginController {
     @FXML
     void initialize() {
 
+        authenticateButton.setOnAction(actionEvent -> {
+
+        });
+    }
+
+    @Override
+    public void start(Stage primaryStage) throws Exception {
+        Parent root = FXMLLoader.load(getClass().getResource("/view/login.fxml"));
+        primaryStage.setTitle("Ghana Electoral Management System");
+        primaryStage.setScene(new Scene(root, 700, 400));
+
+        primaryStage.show();
+    }
+
+    public static void main(String[] args) {
+        launch(args);
     }
 }
